+++ sites/funcampus/src/backend/funcampus/urls.py
@@ -29,6 +29,7 @@ urlpatterns = [
         include([*search_urlpatterns, *plugins_urlpatterns]),
     ),
     path(r"", include("filer.server.urls")),
+    path(r"django-check-seo/", include("django_check_seo.urls")),
 ]
 
 urlpatterns += i18n_patterns(